requirements-test.txt
-r requirements.txt

black
flake8
pylint
mypy
pandas-stubs
